+++ includes/read_input/termcap/init_deinit.h
@@ -1,0 +1,10 @@
+#ifndef INIT_DEINIT_H
+# define INIT_DEINIT_H
+
+# include "read_input/editor/editor_struct.h"
+
+t_term		*init_term();
+void		ft_start_termcaps(void);
+void		ft_close_termcaps(void);
+
+#endif
