+++ srcs/read_input/edit_input.c
@@ -3,6 +3,7 @@
 # include "abstract_list.h"
 # include "read_input/editor/editor.h"
 # include "read_input/event_callbacks/event_callback_def.h"
+# include "read_input/termcap/init_deinit.h"
 
 char *edit_input()
 {
